+++ CHANGELOG.md
@@ -7,4 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [PR-5](https://github.com/itk-dev/tidy-feedback/pull/5)
+  Add styling. Make region and form draggable
+
 [Unreleased]: https://github.com/itk-dev/tidy_feedback